Cache contest ranking
We have 2 things that we can cache:
    (1) The data that we will use to render the ranking table
    (2) The rendered ranking table

Although the first option is pretty is easy to cache, the second one is hard because:
    - If (1) has been updated, we need to update (2) as well

Because I can't find a way to tell whether (1) has been updated or not,
so I save the last timestamp when (1) updated to the cache.
That timestamp will become a cache key for the (2). By doing that,
whenever (1) updated, (2) will use a new cache key.