Allow disable view testcase status
In many official contests, we don't want contestants to be able
to view the per testcase status.

Now, for each problem, we can choose whether we want that user to testcase status.

Also, currently user cannot view the checker feedback. By this commit, we can specify
whether the user can view it or not.