Avoid waiting when only one judge can handle a submission
Before, if there are multiple judges, a submission is low priority, and
only one judge is available, we don't schedule. This is a problem if
only one judge can handle that submission.

We handle the case by delaying rejudges only when there are at least two
judges that can handle the submission, of which only one is available.